Allow separators after implicit release numbers
The PEP 440 regex allows separators between tags and release numbers,
even if those numbers are implicit such as `1.dev.` or `1.post..dev`.

For cases like `1.post.dev` it is not clear whether this should be parsed
as [1][.post][.dev] or [1][.post.][dev]. The former case was chosen as
it is closer to the normalized format.